direct user to job details when a job card is clicked
